Navigation links reseted to "#" moved simplepage.module.styl from /components to /pages and modified the links to import the styles in Datenschutz and Impressum
