+++ src/pages/datenschutz.js
@@ -2,12 +2,12 @@ import React from "react"
 
 import Layout from "../components/layout"
 import SEO from "../components/seo"
-import styles from "../components/simplepage.module.styl"
+import styles from "../simplepage.module.styl"
 
 const Datenschutz = () => (
   <Layout>
     <SEO title="Datenschutz" />
-    <h1 className={styles.headerMain}>Lorem ipsum </h1>
+    <h1>Lorem ipsum </h1>
     <p>
       Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy
       eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam
