+++ src/pages/impressum.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { Link } from "gatsby"
-import styles from "../components/simplepage.module.styl"
+import styles from "../simplepage.module.styl"
 
 import Layout from "../components/layout"
 import SEO from "../components/seo"
@@ -8,7 +8,7 @@ import SEO from "../components/seo"
 const Impressum = () => (
   <Layout>
     <SEO title="Impressum" />
-    <h1 className={styles.headerMain}>Impressum</h1>
+    <h1>Impressum</h1>
     <Link to="/">Go back to the homepage</Link>
   </Layout>
 )
